[Refactor] - DetailVC -> View, VC로 분리
